move walletconnect constructor to separate module
